REFACTOR

+++ src/main.c
@@ -100,14 +100,14 @@ int init_window(t_mrt *data)
     data->mlx = mlx_init(data->cam.image_width, data->cam.image_height, WINDOW_TITLE, true);
 	if (data->mlx == NULL)
 		return (FALSE);
-	data->image = mlx_new_image(data->mlx, data->cam.image_width, data->cam.image_height);
+	data->image = mlx_new_image(data->mlx, (uint32_t)data->cam.image_width, (uint32_t)data->cam.image_height);
 	if (!(data->image))
 	{
 		mlx_close_window(data->mlx);
 		ft_printf("%s\n", mlx_strerror(mlx_errno));
 		
 	}
-	ft_memset(data->image->pixels, 0, data->cam.image_width * data->cam.image_height * BPP);
+	ft_memset(data->image->pixels, (int)0, (size_t)data->cam.image_width * (size_t)data->cam.image_height * (size_t)BPP);
 	if (mlx_image_to_window(data->mlx, data->image, 0, 0) == -1)
 	{
 		mlx_close_window(data->mlx);
@@ -151,7 +151,7 @@ void	_resize_hook(int new_width, int new_height, void *params)
 	data->cam.image_width = new_width;
 	data->cam.image_height = new_height;
 	update_cam(&data->cam, new_width, new_height);
-	mlx_resize_image(data->image, new_width, new_height);
+	mlx_resize_image(data->image, (uint32_t)new_width, (uint32_t)new_height);
 	data->needs_render = true;
 }
